# EKYCBackendService/interfaces/openapi.yaml
openapi: 3.0.3
info:
  title: EKYC Backend API
  version: 1.0.0
  description: API for Registration/Login validation and Bank details management with GxP audit and RBAC.
servers:
  - url: http://localhost:3000
tags:
  - name: Auth
  - name: Bank
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register user
      x-traceability:
        requirements: ['REQ-VAL-001','REQ-AUTH-001','REQ-SEC-001']
        stories:
          - 'As a user, I want clear validation and guidance during registration.'
        validation_protocols: ['VP-VAL-001','VP-SEC-001']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, maxLength: 50 }
                mobile: { type: string, pattern: '^\d{10}$' }
                password: { type: string, minLength: 8 }
      responses:
        '201':
          description: Created
        '400':
          description: Validation error
        '409':
          description: Duplicate account
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login via email or mobile
      x-traceability:
        requirements: ['REQ-AUTH-001','REQ-VAL-001','REQ-SEC-001']
        stories:
          - 'As a registered user, I want to securely log in.'
        validation_protocols: ['VP-VAL-001','VP-SEC-001']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                identifier: { type: string, description: Email or mobile }
                password: { type: string }
      responses:
        '200':
          description: Token
        '401':
          description: Invalid credentials
  /api/bank-details:
    get:
      tags: [Bank]
      summary: Get bank details
      x-traceability:
        requirements: ['REQ-SEC-ACL-001','REQ-SEC-001']
        validation_protocols: ['VP-SEC-001']
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Bank details or null
        '401':
          description: Unauthorized
    post:
      tags: [Bank]
      summary: Create/Update bank details (requires e-sign binding)
      description: |
        Performs a critical update of bank details. Requires reasonForChange and e-sign binding.
        If `critical` is true, the user must re-authenticate by confirming password.
      x-traceability:
        requirements: ['REQ-BANK-001','REQ-BANK-IFSC-001','REQ-ESIGN-001','REQ-SEC-001','REQ-SEC-ACL-001']
        stories:
          - 'As a user, I want to enter my bank account twice and validate IFSC; provide e-sign and reason.'
        validation_protocols: ['VP-BANK-001','VP-ESIGN-001','VP-SEC-001']
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accountNumber, confirmAccountNumber, ifsc, fullNameForESign, agreeESign, reasonForChange]
              properties:
                accountNumber: { type: string, minLength: 8, maxLength: 20 }
                confirmAccountNumber: { type: string, minLength: 8, maxLength: 20 }
                ifsc: { type: string, pattern: '^[A-Z]{4}0[A-Z0-9]{6}$' }
                fullNameForESign: { type: string, minLength: 1 }
                agreeESign: { type: boolean }
                reasonForChange: { type: string, minLength: 1, maxLength: 250 }
                critical: { type: boolean, description: 'If true, requires re-authentication' }
                reauthPassword: { type: string, description: 'Password confirmation required when critical=true' }
      responses:
        '200':
          description: Saved
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (missing permission)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT